Update CodeAnalysis package to V3
Also add TestsRuleset variant to UnitTests project
