+++ .editorconfig
@@ -50,3 +50,4 @@ end_of_line = lf
 indent_size = 4
 insert_final_newline = true
 charset = utf-8-bom
+
