+++ Neolution.DotNet.Console.Sample/Commands/Echo/EchoCommand.cs
@@ -36,7 +36,7 @@
                 throw new ArgumentNullException(nameof(options));
             }
 
-            this.logger.LogInformation(options.Message);
+            this.logger.LogInformation("Called with message: {Message}", options.Message);
         }
     }
 }
