+++ Neolution.DotNet.Console.SampleAsync/Neolution.DotNet.Console.SampleAsync.csproj
@@ -11,7 +11,7 @@
     <PackageReference Include="AWS.Logger.NLog" Version="3.0.0" />
     <PackageReference Include="CommandLineParser" Version="2.8.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="6.0.1" />
-    <PackageReference Include="Neolution.CodeAnalysis" Version="2.5.8">
+    <PackageReference Include="Neolution.CodeAnalysis" Version="3.0.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
