+++ Neolution.DotNet.Console.UnitTests/Common/Stubs/IScopedServiceStub.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.DotNet.Console.UnitTests.Common.Stubs
+{
+    /// <summary>
+    /// Stub for a generic scoped service.
+    /// </summary>
+    public interface IScopedServiceStub
+    {
+        /// <summary>
+        /// Does something.
+        /// </summary>
+        void DoSomething();
+    }
+}
