+++ Neolution.DotNet.Console.UnitTests/Common/Stubs/ISingletonServiceStub.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.DotNet.Console.UnitTests.Common.Stubs
+{
+    /// <summary>
+    /// Stub for a generic singleton service.
+    /// </summary>
+    public interface ISingletonServiceStub
+    {
+        /// <summary>
+        /// Does something.
+        /// </summary>
+        void DoSomething();
+    }
+}
