+++ Neolution.DotNet.Console.UnitTests/Common/Stubs/ITransientServiceStub.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.DotNet.Console.UnitTests.Common.Stubs
+{
+    /// <summary>
+    /// Stub for a generic transient service.
+    /// </summary>
+    public interface ITransientServiceStub
+    {
+        /// <summary>
+        /// Does something.
+        /// </summary>
+        void DoSomething();
+    }
+}
