--- Neolution.DotNet.Console.UnitTests/Common/Stubs/ServiceStubs.cs
@@ -1,25 +1,0 @@
-﻿namespace Neolution.DotNet.Console.UnitTests.Common.Stubs;
-
-public interface ITransientServiceStub
-{
-}
-
-public class TransientServiceStub : ITransientServiceStub
-{
-}
-
-public interface IScopedServiceStub
-{
-}
-
-public class ScopedServiceStub : IScopedServiceStub
-{
-}
-
-public interface ISingletonServiceStub
-{
-}
-
-public class SingletonServiceStub : ISingletonServiceStub
-{
-}
