+++ Neolution.DotNet.Console.UnitTests/Common/UnitTestsConstants.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.DotNet.Console.UnitTests.Common
+{
+    /// <summary>
+    /// Constant values to use in Tests
+    /// </summary>
+    internal static class UnitTestsConstants
+    {
+        /// <summary>
+        /// The command name
+        /// </summary>
+        internal const string InjectServicesWithVariousLifetimes = "InjectServicesWithVariousLifetimes";
+    }
+}
