+++ Neolution.DotNet.Console.UnitTests/ConsoleAppStartup/Stubs/InjectServicesWithVariousLifetimesOptionsStub.cs
@@ -1,16 +1,13 @@
 ﻿namespace Neolution.DotNet.Console.UnitTests.ConsoleAppStartup.Stubs
 {
     using CommandLine;
+    using Neolution.DotNet.Console.UnitTests.Common;
 
     /// <summary>
     /// The options stub for the <see cref="InjectServicesWithVariousLifetimesCommandStub"/>
     /// </summary>
-    [Verb(CommandName)]
+    [Verb(UnitTestsConstants.InjectServicesWithVariousLifetimes)]
     public class InjectServicesWithVariousLifetimesOptionsStub
     {
-        /// <summary>
-        /// The command name
-        /// </summary>
-        public const string CommandName = "InjectServicesWithVariousLifetimes";
     }
 }
